+++ src/cartesian/CartesianAxis.tsx
@@ -450,8 +450,8 @@ export class CartesianAxis extends Component<Props, IState> {
   /**
    * render the ticks
    * @param {Array} ticks The ticks to actually render (overrides what was passed in props)
-   * @param {string} fontSize The size of the font used in labels
-   * @param {string} letterSpacing The space between letters of the font used in labels
+   * @param {string} fontSize Fontsize to consider for tick spacing
+   * @param {string} letterSpacing Letterspacing to consider for tick spacing
    * @return {ReactComponent} renderedTicks
    */
   renderTicks(ticks: CartesianTickItem[], fontSize: string, letterSpacing: string) {