+++ src/chart/generateCategoricalChart.tsx
@@ -539,7 +539,7 @@ const createDefaultState = (props: CategoricalChartProps): CategoricalChartState
   const brushItem = findChildByType(children, Brush.displayName);
   const startIndex = (brushItem && brushItem.props && brushItem.props.startIndex) || 0;
   const endIndex =
-    (brushItem && brushItem.props && brushItem.props.endIndex) || (props.data && props.data.length - 1) || 0;
+    brushItem?.props?.endIndex !== undefined ? brushItem?.props?.endIndex : (props.data && props.data.length - 1) || 0;
 
   return {
     chartX: 0,
@@ -961,7 +961,6 @@ export const generateCategoricalChart = ({
       this.state = {};
     }
 
-    /* eslint-disable  react/no-did-mount-set-state */
     componentDidMount() {
       if (!_.isNil(this.props.syncId)) {
         this.addListener();
@@ -1443,6 +1442,8 @@ export const generateCategoricalChart = ({
 
         const handler = event;
 
+        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+        // @ts-ignore
         handler(mouse, e);
       }
     };
@@ -1675,7 +1676,7 @@ export const generateCategoricalChart = ({
     renderPolarAxis = (element: any, displayName: string, index: number) => {
       const axisType = _.get(element, 'type.axisType');
       const axisMap = _.get(this.state, `${axisType}Map`);
-      const axisOption = axisMap[element.props[`${axisType}Id`]];
+      const axisOption: BaseAxisProps | undefined = axisMap && axisMap[element.props[`${axisType}Id`]];
 
       return cloneElement(element, {
         ...axisOption,
@@ -2137,6 +2138,7 @@ export const generateCategoricalChart = ({
           ref={node => {
             this.container = node;
           }}
+          role="region"
         >
           <Surface {...attrs} width={width} height={height} title={title} desc={desc}>
             {this.renderClipPath()}